vulkan-sdk-option (#143)
Provides an option to build against the Vulkan SDK or the
Vulkan-Headers repo.
The default is to build against the Vulkan-Headers repo.
To build against the SDK, configure with `-D gvk-Vulkan-SDK_EXTRACT=ON`
The target SDK is still downloaded during the build for convenience.

NOTE : Some functionality may not behave properly without
`VK_LAYER_INTEL_gvk_unique_handles` which will be provided in a follow
on PR.

NOTE : While the vanilla default is now to build against Vulkan-Headers,
internal builds should be using `-C internal/build-options.cmake` which
will enable SDK extraction...this will likely change in the near future,
but CI needs an update.

TODO : Need to setup `GVK_Daily` pipeline to validate building against
both.

TODO : Maybe an option to install the SDK during the build